+++ ReleaseNotes/release_notes.md
@@ -1,9 +1,14 @@
+# v3.6.4.0 
+
 ## New features
 
    1. `Ostap::Math::Sigmoid` : add more sigmoid types; normalize all sigmoids to have the same slope at the x0
    1. add methods to add (constant) `ROOT.RooCategory` to dataset &extend the test
    1. remove the upper limit of nunmber of used cpus 
    1. re-write `summary_graph`, make it more robust& more intuitive
+   1. `Ostap::StatVar` : extend the error message for `Ostap::Assert`
+   1. Ostap::AddVar`  : allow to add RooReal
+   1. add `|`into list of symbols not allowed for good names
    
 ## Bug fixes